it's 4:02 AM!

+++ JBotSim/src/simulation/environment/Environment.java
@@ -147,7 +147,7 @@ public abstract class Environment implements KeyListener, Serializable {
 		teleported.add(physicalObject);
 	}
 
-	public void updateCollisions(int time) {
+	public void updateCollisions(double time) {
 		// updateRobotCloseObjects(time);
 		collisionManager.handleCollisions(this, time);
 	}
