+++ JBotSim/src/simulation/robot/Robot.java
@@ -243,7 +243,7 @@ public class Robot extends MovableObject {
 	 * @param simulationStep the number of the current simulation step.
 	 * @param timeDelta      the time (in virtual seconds) between calls to this method. 
 	 */
-	public void updateActuators(int simulationStep, double timeDelta) {	
+	public void updateActuators(double simulationStep, double timeDelta) {	
 		position.set(
 				position.getX() + timeDelta * (leftWheelSpeed + rightWheelSpeed) / 2.0 * Math.cos(orientation),
 				position.getY() + timeDelta * (leftWheelSpeed + rightWheelSpeed) / 2.0 * Math.sin(orientation));
